fix(#27): new Kit item data is not saved
This started when we switched to use Mojang's Encoder tech in the 1.20.5
upgrade.

I maintain that that CODEC/Encoder suite is one of the most opaque and
incomprehensible systems I've ever had the displeasure of looking at.

At it's core, these encoders seem to be facilitating immutability-focused
operations. If I'd realized that, probably could have picked up on this issue
earlier. (Fundamentally, the code was expecting the `encode` operation to modify
an `NbtCompound`, as the old `writeNbt` would, but it returns a new `NbtElement`
instead. The solution is to write that `NbtElement`, instead of our (now
"prefix") `nbtCompound`.